MarketHive/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
dj-database-url>=2.1.0
stripe>=7.0.0
boto3>=1.34.0
django-storages>=1.14.0
gunicorn>=21.2.0
whitenoise>=6.6.0
bcrypt>=4.1.0
Pillow>=10.0.0
